Renamed TestingAppStateInitializer and added Javadocs
